Moves Starships out of directory into its own file
